move i18next translation resource to helper
